refactor: metrics
