+++ iroh-gateway/src/client.rs
@@ -1,15 +1,15 @@
-use std::sync::Arc;
-
 use axum::body::StreamBody;
+use iroh_metrics::gateway::Metrics;
 use iroh_resolver::resolver::CidOrDomain;
 use iroh_resolver::resolver::Metadata;
 use iroh_resolver::resolver::OutPrettyReader;
 use iroh_resolver::resolver::Resolver;
+use iroh_resolver::resolver::Source;
+use prometheus_client::registry::Registry;
 use tokio_util::io::ReaderStream;
 use tracing::info;
 
 use crate::core::GetParams;
-use crate::core::State;
 use crate::response::ResponseFormat;
 
 #[derive(Debug)]
@@ -20,9 +20,9 @@ pub struct Client {
 pub type PrettyStreamBody = StreamBody<ReaderStream<OutPrettyReader<iroh_rpc_client::Client>>>;
 
 impl Client {
-    pub fn new(rpc_client: &iroh_rpc_client::Client) -> Self {
+    pub fn new(rpc_client: &iroh_rpc_client::Client, registry: &mut Registry) -> Self {
         Self {
-            resolver: Resolver::new(rpc_client.clone()),
+            resolver: Resolver::new(rpc_client.clone(), registry),
         }
     }
 
@@ -32,26 +32,28 @@ impl Client {
         path: iroh_resolver::resolver::Path,
         rpc_client: &iroh_rpc_client::Client,
         start_time: std::time::Instant,
-        state: Arc<State>,
+        metrics: &Metrics,
     ) -> Result<(PrettyStreamBody, Metadata), String> {
         info!("get file {}", path);
-        state.metrics.cache_miss.inc();
-        // todo(arqu): this is wrong but currently don't have access to the data stream
-        state
-            .metrics
-            .ttf_block
-            .set(start_time.elapsed().as_millis() as u64);
-        state
-            .metrics
-            .hist_ttfb
-            .observe(start_time.elapsed().as_millis() as f64);
         let res = self
             .resolver
             .resolve(path)
             .await
             .map_err(|e| e.to_string())?;
+        metrics
+            .ttf_block
+            .set(start_time.elapsed().as_millis() as u64);
         let metadata = res.metadata().clone();
-        let reader = res.pretty(rpc_client.clone());
+        if metadata.source == Source::Bitswap {
+            metrics
+                .hist_ttfb
+                .observe(start_time.elapsed().as_millis() as f64);
+        } else {
+            metrics
+                .hist_ttfb_cached
+                .observe(start_time.elapsed().as_millis() as f64);
+        }
+        let reader = res.pretty(rpc_client.clone(), metrics.clone(), start_time);
         let stream = ReaderStream::new(reader);
         let body = StreamBody::new(stream);
