+++ iroh-resolver/Cargo.toml
@@ -21,6 +21,8 @@ tracing = "0.1.34"
 async-trait = "0.1.53"
 async-recursion = "1.0.0"
 trust-dns-resolver = { version = "0.21.2", features = ["tokio-runtime"] }
+prometheus-client = "0.16.0"
+iroh-metrics = { path = "../iroh-metrics" }
 
 [dev-dependencies]
 criterion = { version = "0.3.5", features = ["async_tokio"] }
